add useCanAccess
